Move core api code into sub-module
